refactor: 修正代码中的一些问题
1. 添加pagehelper插件依赖
2. 将PageResult的total成员变量及其相关total变量的数据类型从Integer提升为Long
3. 修正代码变量、方法命名问题

+++ qk-management/src/main/java/com/qk/management/mapper/ActivityMapper.java
@@ -14,7 +14,7 @@ import java.util.List;
  */
 @Mapper
 public interface ActivityMapper {
-    Integer count(Integer channel, Integer type, Integer status);
+    Long count(Integer channel, Integer type, Integer status);
 
     List<Activity> select(Integer channel, Integer type, Integer status, Integer offset, Integer pageSize);
