+++ qk-management/src/main/java/com/qk/management/mapper/RoleMapper.java
@@ -14,7 +14,7 @@ import java.util.List;
  */
 @Mapper
 public interface RoleMapper {
-    Integer count(String name, String label);
+    Long count(String name, String label);
 
     List<Role> select(String name, String label, Integer offset, Integer pageSize);
